Get an instance of the interactive shiny occupation coding app.

Printing the returned instance or returning it without saving it in a variable will start the app.

Usage

app(
  questionnaire = questionnaire_web_survey(),
  app_settings = create_app_settings(save_to_file = TRUE),
  css_file = NULL,
  resource_dir = system.file("www", package = "occupationMeasurement"),
  ...
)

Arguments

questionnaire

The questionnaire to load. (Defaults to the questionnaire returned by questionnaire_web_survey().)

app_settings

The app_settings to use. Check the documentation for create_app_settings to learn about the options.

css_file

Path to a CSS file to be included in the app.

resource_dir

From which directory to static files e.g. styles. If you want to load additional resources from outside the package, you should rather do so with shiny::addResourcePath rather than with this parameter.

...

Any additional parameters will be forwarded to shiny::shinyApp().

Value

A shiny app instance.

See also

Examples

if (FALSE) { # \dontrun{
app_instance <- app(
  app_settings = create_app_settings(
    # Important to save results from the app
    save_to_file = TRUE
  )
)

# Start the app
if (interactive()) {
  app_instance
}
} # }