Convert suggestion from one suggestion format into the other

Convert suggestion from one suggestion format into the other

Usage

convert_suggestions(suggestions, from, to, suggestion_type_options = list())

Arguments

suggestions

A data.table of suggestions as returned by one of the prediction algorithms e.g. algo_similarity_based_reasoning().

from

The current suggestion_type in which suggestions are passed.

to

The suggestion_type in which suggestions should be.

suggestion_type_options

A list with options for generating suggestions. Supported options: - datasets: Pass specific datasets to be used whenn adding information to predictions e.g. use a specific version of the kldb or auxco. Supported datasets are: "auxco-1.2.x", "kldb-2010". By default the datasets bundled with this package are used.

Value

A data.table of suggestions in the "to" format