Clean & Load KldB 2010 dataset.

Use load_kldb_raw() to load the whole dataset.

Usage

load_kldb_raw(
  cache_dir = getOption("occupationMeasurement.cache_dir", tempdir())
)

load_kldb(cache_dir = getOption("occupationMeasurement.cache_dir", tempdir()))

Arguments

cache_dir

The path to the directory where the downloaded data should be stored. We recommend setting this to "cache" to store data in the working directory. This will prevent reloading the data time and time again. This can be set globally via options(occupationMeasurement.cache_dir = "cache").

Value

A cleaned / slimmed version of the KldB 2010.

Details

Source: https://www.klassifikationsserver.de/klassService/index.jsp?variant=kldb2010

More information on the KldB 2010: https://statistik.arbeitsagentur.de/DE/Navigation/Grundlagen/Klassifikationen/Klassifikation-der-Berufe/KldB2010-Fassung2020/KldB2010-Fassung2020-Nav.html The KldB 2010 has been revised in 2020. These changes have not been implemented here yet.

Functions

load_kldb_raw(): Load raw KldB 2010 dataset.

Examples

# We recommend using a non-temporary directory for caching, so data is
# downloaded only once and not time and time again
cache_dir <- tempdir()
# Note: The dataset will be downloaded from the internet
# Load the cleaned dataset
load_kldb(cache_dir = cache_dir)
#> [1] "Using a modified version of the KldB 2010.\nPlease mind the terms of use of the original KldB dataset (German):\n\n    (c) Statistik der Bundesagentur für Arbeit\n    Sie können Informationen speichern, (auch auszugsweise) mit Quellenangabe\n    weitergeben, vervielfältigen und verbreiten. Die Inhalte dürfen nicht\n    verändert oder verfälscht werden. Eigene Berechnungen sind erlaubt, jedoch\n    als solche kenntlich zu machen. Im Falle einer Zugänglichmachung im\n    Internet soll dies in Form einer Verlinkung auf die Homepage der Statistik\n    der Bundesagentur für Arbeit erfolgen. Die Nutzung der Inhalte für\n    gewerbliche Zwecke, ausgenommen Presse, Rundfunk und Fernsehen und\n    wissenschaftliche Publikationen, bedarf der Genehmigung durch die Statistik\n    der Bundesagentur für Arbeit.\n  "
#>       kldb_id level                                                     label
#>    1:       1     1                                                      <NA>
#>    2:      11     2                                                      <NA>
#>    3:     111     3                                                      <NA>
#>    4:    1110     4                     Landwirtschaft (ohne Spezialisierung)
#>    5:   11101     5                     Landwirtschaft (ohne Spezialisierung)
#>   ---                                                                        
#> 2173:    0130     4                              Unteroffiziere ohne Portepee
#> 2174:   01302     5                              Unteroffiziere ohne Portepee
#> 2175:     014     3                                                      <NA>
#> 2176:    0140     4 Angehörige der regulären Streitkräfte in sonstigen Rängen
#> 2177:   01402     5 Angehörige der regulären Streitkräfte in sonstigen Rängen
#>                                                                                                                                                                                                                                                                                                                                                                                                                                  description
#>    1:                                                                                                                                                                                                                                                                                                                                                                                                                                       
#>    2:                                                                                                                                                                                                                                                       Inhalt: Diese Systematikposition umfasst die Berufe in der Land-, Tier-, Pferde- und Fischwirtschaft, Tierpflege, Weinbau sowie Forst- und Jagdwirtschaft und Landschaftspflege.
#>    3:                                                                                                                                                                                                                                   Inhalt: Angehörige dieser Berufe übernehmen Aufgaben in der Landwirtschaft, in der Landtechnik oder im landwirtschaftlich-technischen Labor oder sind als landwirtschaftliche Sachverständige tätig.
#>    4:                                                                                                                                         Inhalt: Angehörige dieser Berufe sind in der Herstellung landwirtschaftlicher Produkte tätig. Sie arbeiten in der Pflanzen- und Tierproduktion, bedienen Betriebsanlagen, vermarkten Produkte, beraten landwirtschaftliche Betriebe oder planen den Einsatz von Technik in der Landwirtschaft.
#>    5:                                      Inhalt: Diese Systematikposition umfasst alle Berufe in der Landwirtschaft, deren Tätigkeiten in der Regel keine speziellen Fachkenntnisse erfordern. Angehörige dieser Berufe führen in landwirtschaftlichen Betrieben nach Anweisung einfachere oder zuarbeitende Tätigkeiten aus. Je nach Einsatzgebiet bearbeiten sie die Erde, versorgen Nutztiere und übernehmen Pflanz- und Erntearbeiten.
#>   ---                                                                                                                                                                                                                                                                                                                                                                                                                                       
#> 2173:                                                                                                                                                                         Inhalt: Angehörige dieser Berufe erfüllen bei der Bundeswehr im Unteroffiziersrang ohne Portepee zumeist Aufgaben, die denen des erlernten zivilen Berufs entsprechen. Sie sind im allgemeinen Fachdienst tätig, im Sanitätsdienst oder im Militärmusikdienst.
#> 2174:                   Inhalt: Diese Systematikposition umfasst alle Militärangehörigen der Bundeswehr, deren Tätigkeiten fundierte fachliche Kenntnisse und Fertigkeiten erfordern. Angehörige dieser Berufe erfüllen im Unteroffiziersrang ohne Portepee bei der Bundeswehr zumeist Aufgaben, die denen des erlernten zivilen Berufs entsprechen. Sie sind im allgemeinen Fachdienst tätig, im Sanitätsdienst oder im Militärmusikdienst.
#> 2175:                                                                                                                                                                                                                                       Inhalt: Angehörige dieser Berufe erfüllen bei der Bundeswehr militärspezifische Funktionen in Mannschaftsdienstgraden, für die eine zivilberufliche Ausbildung nicht unbedingt erforderlich ist.
#> 2176:                                                                                                                                                 Inhalt: Angehörige dieser Berufe erfüllen bei der Bundeswehr militärspezifische Funktionen in Mannschaftsdienstgraden, für die eine zivilberufliche Ausbildung nicht erforderlich ist. In einzelnen Verwendungen setzen sie jedoch auch erworbene zivilberufliche Qualifikationen ein.
#> 2177: Inhalt: Diese Systematikposition umfasst alle Militärangehörigen der regulären Streitkräfte in Mannschaftsdienstgraden, deren Tätigkeiten fundierte fachliche Kenntnisse und -fertigkeiten erfordern. Angehörige dieser Berufe erfüllen militärspezifische Funktionen, für die eine zivilberufliche Ausbildung nicht erforderlich ist. In einzelnen Verwendungen setzen sie jedoch auch erworbene zivilberufliche Qualifikationen ein.
#>                                                                                                                                                                                                                                                                                                                                                                           excludes
#>    1:                                                                                                                                                                                                                                                                                                                                                                             
#>    2:                                                                                                                                                                                                                                                                                                                                                                             
#>    3:                                                                                                                                                                                                                       Nicht einzubeziehende Positionen: 112 Tierwirtschaft 114 Fischwirtschaft 115 Tierpflege 117 Forst- und Jagdwirtschaft, Landschaftspflege 121 Gartenbau
#>    4: Nicht einzubeziehende Positionen: 1121Berufe in der Nutztierhaltung (außer Geflügelhaltung) 1130Berufe in der Pferdewirtschaft (ohne Spezialisierung) 1151Berufe in der Nutztierpflege 1160Berufe im Weinbau 1171Berufe in der Forstwirtschaft 1212Berufe in Baumschule, Staudengärtnerei und Zierpflanzenbau 5251Führer/innen von land- und forstwirtschaftlichen Maschinen
#>    5:                                                                                                                                                                                                                       Nicht einzubeziehende Berufe: Tierwirtschaftshelfer/in (11211) Fischereihelfer/in (11401) Forstwirtschaftshelfer/in (11711) Gartenbauhelfer/in (12101)
#>   ---                                                                                                                                                                                                                                                                                                                                                                             
#> 2173:                                                                                                                                                                                                                             Nicht einzubeziehende Positionen: 0110 Offiziere 0120 Unteroffiziere mit Portepee 0140 Angehörige der regulären Streitkräfte in sonstigen Rängen
#> 2174:                                                                                                                                                                                                                                                           Nicht einzubeziehende Berufe: Fähnrich (01203) Stabsfeldwebel (01203) Oberstabsfeldwebel (01203) Feldwebel (01203)
#> 2175:                                                                                                                                                                                                                                                             Nicht einzubeziehende Positionen: 011 Offiziere 012 Unteroffiziere mit Portepee 013 Unteroffiziere ohne Portepee
#> 2176:                                                                                                                                                                                                                                                          Nicht einzubeziehende Positionen: 0110 Offiziere 0120 Unteroffiziere mit Portepee 0130 Unteroffiziere ohne Portepee
#> 2177:                                                                                                                                                                                                                                                                                                                                                                             
#>                                                                                 title
#>    1:                                  Land-, Forst- und Tierwirtschaft und Gartenbau
#>    2:                                         Land-, Tier- und Forstwirtschaftsberufe
#>    3:                                                                  Landwirtschaft
#>    4:                             Berufe in der Landwirtschaft (ohne Spezialisierung)
#>    5: Berufe in der Landwirtschaft (ohne Spezialisierung) - Helfer-/Anlerntätigkeiten
#>   ---                                                                                
#> 2173:                                                    Unteroffiziere ohne Portepee
#> 2174:                                                    Unteroffiziere ohne Portepee
#> 2175:                       Angehörige der regulären Streitkräfte in sonstigen Rängen
#> 2176:                       Angehörige der regulären Streitkräfte in sonstigen Rängen
#> 2177:                       Angehörige der regulären Streitkräfte in sonstigen Rängen
# Load the raw dataset
load_kldb_raw(cache_dir = cache_dir)
#>       Schlüssel KldB 2010 Ebene
#>    1:                   1     1
#>    2:                  11     2
#>    3:                 111     3
#>    4:                1110     4
#>    5:               11101     5
#>   ---                          
#> 2173:                0130     4
#> 2174:               01302     5
#> 2175:                 014     3
#> 2176:                0140     4
#> 2177:               01402     5
#>                                                                                 Titel
#>    1:                                  Land-, Forst- und Tierwirtschaft und Gartenbau
#>    2:                                         Land-, Tier- und Forstwirtschaftsberufe
#>    3:                                                                  Landwirtschaft
#>    4:                             Berufe in der Landwirtschaft (ohne Spezialisierung)
#>    5: Berufe in der Landwirtschaft (ohne Spezialisierung) - Helfer-/Anlerntätigkeiten
#>   ---                                                                                
#> 2173:                                                    Unteroffiziere ohne Portepee
#> 2174:                                                    Unteroffiziere ohne Portepee
#> 2175:                       Angehörige der regulären Streitkräfte in sonstigen Rängen
#> 2176:                       Angehörige der regulären Streitkräfte in sonstigen Rängen
#> 2177:                       Angehörige der regulären Streitkräfte in sonstigen Rängen
#>       Erläuterungstitel                                Kurztitel Mitteltitel
#>    1:                   Land-, Forst-, Tierwirtschaft, Gartenbau          NA
#>    2:                       Land-, Tier-, Forstwirtschaftsberufe          NA
#>    3:                                             Landwirtschaft          NA
#>    4:                        Berufe in der Landwirtschaft (o.S.)          NA
#>    5:                             Landwirtschaft (o.S.) - Helfer          NA
#>   ---                                                                       
#> 2173:                               Unteroffiziere ohne Portepee          NA
#> 2174:                               Unteroffiziere ohne Portepee          NA
#> 2175:                   Angeh. reguläre Streitkräfte sonst.Ränge          NA
#> 2176:                   Angeh. reguläre Streitkräfte sonst.Ränge          NA
#> 2177:                   Angeh. reguläre Streitkräfte sonst.Ränge          NA
#>       Langtitel
#>    1:        NA
#>    2:        NA
#>    3:        NA
#>    4:        NA
#>    5:        NA
#>   ---          
#> 2173:        NA
#> 2174:        NA
#> 2175:        NA
#> 2176:        NA
#> 2177:        NA
#>                                                                                                                                                                                                                                                                                                                                                                                                                       Allgemeine Bemerkungen
#>    1:                                                                                                                                                                                                                                                                                                                                                                                                                                       
#>    2:                                                                                                                                                                                                                                                       Inhalt: Diese Systematikposition umfasst die Berufe in der Land-, Tier-, Pferde- und Fischwirtschaft, Tierpflege, Weinbau sowie Forst- und Jagdwirtschaft und Landschaftspflege.
#>    3:                                                                                                                                                                                                                                   Inhalt: Angehörige dieser Berufe übernehmen Aufgaben in der Landwirtschaft, in der Landtechnik oder im landwirtschaftlich-technischen Labor oder sind als landwirtschaftliche Sachverständige tätig.
#>    4:                                                                                                                                         Inhalt: Angehörige dieser Berufe sind in der Herstellung landwirtschaftlicher Produkte tätig. Sie arbeiten in der Pflanzen- und Tierproduktion, bedienen Betriebsanlagen, vermarkten Produkte, beraten landwirtschaftliche Betriebe oder planen den Einsatz von Technik in der Landwirtschaft.
#>    5:                                      Inhalt: Diese Systematikposition umfasst alle Berufe in der Landwirtschaft, deren Tätigkeiten in der Regel keine speziellen Fachkenntnisse erfordern. Angehörige dieser Berufe führen in landwirtschaftlichen Betrieben nach Anweisung einfachere oder zuarbeitende Tätigkeiten aus. Je nach Einsatzgebiet bearbeiten sie die Erde, versorgen Nutztiere und übernehmen Pflanz- und Erntearbeiten.
#>   ---                                                                                                                                                                                                                                                                                                                                                                                                                                       
#> 2173:                                                                                                                                                                         Inhalt: Angehörige dieser Berufe erfüllen bei der Bundeswehr im Unteroffiziersrang ohne Portepee zumeist Aufgaben, die denen des erlernten zivilen Berufs entsprechen. Sie sind im allgemeinen Fachdienst tätig, im Sanitätsdienst oder im Militärmusikdienst.
#> 2174:                   Inhalt: Diese Systematikposition umfasst alle Militärangehörigen der Bundeswehr, deren Tätigkeiten fundierte fachliche Kenntnisse und Fertigkeiten erfordern. Angehörige dieser Berufe erfüllen im Unteroffiziersrang ohne Portepee bei der Bundeswehr zumeist Aufgaben, die denen des erlernten zivilen Berufs entsprechen. Sie sind im allgemeinen Fachdienst tätig, im Sanitätsdienst oder im Militärmusikdienst.
#> 2175:                                                                                                                                                                                                                                       Inhalt: Angehörige dieser Berufe erfüllen bei der Bundeswehr militärspezifische Funktionen in Mannschaftsdienstgraden, für die eine zivilberufliche Ausbildung nicht unbedingt erforderlich ist.
#> 2176:                                                                                                                                                 Inhalt: Angehörige dieser Berufe erfüllen bei der Bundeswehr militärspezifische Funktionen in Mannschaftsdienstgraden, für die eine zivilberufliche Ausbildung nicht erforderlich ist. In einzelnen Verwendungen setzen sie jedoch auch erworbene zivilberufliche Qualifikationen ein.
#> 2177: Inhalt: Diese Systematikposition umfasst alle Militärangehörigen der regulären Streitkräfte in Mannschaftsdienstgraden, deren Tätigkeiten fundierte fachliche Kenntnisse und -fertigkeiten erfordern. Angehörige dieser Berufe erfüllen militärspezifische Funktionen, für die eine zivilberufliche Ausbildung nicht erforderlich ist. In einzelnen Verwendungen setzen sie jedoch auch erworbene zivilberufliche Qualifikationen ein.
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               Einschlüsse
#>    1:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    
#>    2:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    
#>    3:                                                             Aufgaben, Tätigkeiten, Kenntnisse und Fertigkeiten, üblicherweise: ● die Erzeugung landwirtschaftlicher Produkte planen, organisieren, optimieren und durchführen, die Erzeugnisse vermarkten ● Maschinen und Betriebsanlagen in der Landwirtschaft bedienen und warten, den Einsatz planen und überwachen, Verfahren entwickeln bzw. optimieren ● Analysen von Saatgut, Düngemittel und Bodenproben im landwirtschaftlichen Versuchs- und Untersuchungswesen durchführen  ● die Betriebspolitik (mit-)gestalten, betriebswirtschaftliche Aufgaben erledigen, z.B. Investitionsent-scheidungen treffen, Kosten- und Leistungsrechnung durchführen ● Tätigkeiten in der Landwirtschafts- bzw. Agrarverwaltung durchführen, z.B. die Umsetzung von Förderprogrammen betreuen ● Gutachten, Stellungnahmen und Wertermittlungen zu landwirtschaftlichen Sachverhalten erstellen ● Spezialaufgaben in der Landwirtschaft übernehmen, z.B. im Bereich Direktvermarktung beraten, die Agrarverwaltung EDV-gestützt durchführen
#>    4:                                                                                                                                                                                                                                                                                                                                                                                                                                              Aufgaben, Tätigkeiten, Kenntnisse und Fertigkeiten, üblicherweise: ● Erzeugung pflanzlicher und tierischer Produkte planen, bei Anbau und Pflanzenpflege sowie Tierhaltung praktisch mitarbeiten, Erzeugnisse verkaufen bzw. vermarkten ● landwirtschaftliche Anlagen, Maschinen und Geräte bedienen und warten, Einsatz von landwirtschaft-lichen Anlagen, Maschinen und Geräten steuern ● Aufgaben in der Agrarverwaltung übernehmen, Umsetzung von Förderprogrammen begleiten ● landwirtschaftliche Betriebe und Organisationen beraten ● an der Weiterentwicklung landwirtschaftlicher Produkte und Produktionsverfahren mitwirken
#>    5:                                                                                                                                                                                                                                                                                                                                                                                                         Aufgaben, Tätigkeiten, Kenntnisse und Fertigkeiten, üblicherweise: ● Gras, Heu oder Stroh zusammenrechen und laden ● Pflanzen mit Hilfe von Handwerkzeugen bewässern, ausdünnen und jäten  ● Hilfsarbeiten bei der Ernte durchführen ● Obst, Nüsse, Gemüse und andere Früchte pflücken oder aufsammeln ● Tiere füttern, tränken und Tierunterkünfte reinigen ● Reinigungsarbeiten an landwirtschaftlichen Einrichtungen und Maschinen durchführen ● kleinere Reparaturen an Tierunterkünften und landwirtschaftlichen Einrichtungen durchführen ● Hilfsmittel, Erzeugnisse und sonstige Materialien, z.B. Futtermittel, tierische und pflanzliche Produkte be- und entladen
#>   ---                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    
#> 2173: Aufgaben, Tätigkeiten, Kenntnisse und Fertigkeiten, üblicherweise: ● für den erlernten Zivilberuf typische Aufgaben in allen Uniformträgerbereichen (Heer, Luftwaffe, Marine, Streit-kräftebasis) übernehmen ● aufgrund des Dienstgrades Vorgesetztenfunktionen im Sinne einer allgemeinen Befehlsbefugnis gegenüber den Mannschaften wahrnehmen ● Fachaufgaben im allgemeinen Fachdienst wahrnehmen, z.B. in der Elektronikinstandsetzung, der Luftfahrzeugausrüs-tungstechnik, in der Kfz-Instandsetzung oder im Bereich Verpflegung ● Patienten und Patientinnen betreuen, Blut abnehmen, Verbände anlegen, über Prophylaxe beraten, bei der ärztlichen Behandlung assistieren, Patientenakten führen ● medizinische Instrumente, Geräte und Apparate verwenden, pflegen, desinfizieren und sterilisieren, den Bestand an Medikamenten und anderem Medizinbedarf pflegen ● das Musikrepertoire, je nach musikalischer Richtung des Musikkorps, einstudieren und vortragen, ggf. Tonanlagen aufbauen und Soundchecks durchführen ● gegebenenfalls an Auslandseinsätzen teilnehmen
#> 2174: Aufgaben, Tätigkeiten, Kenntnisse und Fertigkeiten, üblicherweise: ● für den erlernten Zivilberuf typische Aufgaben in allen Uniformträgerbereichen (Heer, Luftwaffe, Marine, Streit-kräftebasis) übernehmen ● aufgrund des Dienstgrades Vorgesetztenfunktionen im Sinne einer allgemeinen Befehlsbefugnis gegenüber den Mannschaften wahrnehmen ● Fachaufgaben im allgemeinen Fachdienst wahrnehmen, z.B. in der Elektronikinstandsetzung, der Luftfahrzeugausrüs-tungstechnik, in der Kfz-Instandsetzung oder im Bereich Verpflegung ● Patienten und Patientinnen betreuen, Blut abnehmen, Verbände anlegen, über Prophylaxe beraten, bei der ärztlichen Behandlung assistieren, Patientenakten führen ● medizinische Instrumente, Geräte und Apparate verwenden, pflegen, desinfizieren und sterilisieren, den Bestand an Medikamenten und anderem Medizinbedarf pflegen ● das Musikrepertoire, je nach musikalischer Richtung des Musikkorps, einstudieren und vortragen, ggf. Tonanlagen aufbauen und Soundchecks durchführen ● gegebenenfalls an Auslandseinsätzen teilnehmen
#> 2175:                                                                                                                                                                                                                                                                                                                     Aufgaben, Tätigkeiten, Kenntnisse und Fertigkeiten, üblicherweise: ● militärspezifische Funktionen erfüllen, z.B. als Fallschirmjäger, Kraftfahrer/in, Flugabwehrraketensoldat/in, Gebirgs-jäger oder Panzergrenadier ● in Verwendungen tätig sein, die zivilberufliche Qualifikationen voraussetzen oder in deren Rahmen eine Qualifikation erworben werden kann, z.B. als Radarelektroniksoldat/in, Raketenelektronikmechaniker/in oder Personalsoldat/in ● technische Geräte und Anlagen bedienen, warten und ggf. reparieren ● als Sanitätssoldat/in oder im Militärmusikdienst tätig sein ● im Stabsdienst auf der Streitkräftebasis arbeiten, z.B. die Planung und Durchführung logistischer Einsätze unterstützen, satellitengestützte Aufklärung nutzen
#> 2176:                                                                                                                               Aufgaben, Tätigkeiten, Kenntnisse und Fertigkeiten, üblicherweise: ● militärspezifische Funktionen erfüllen, z.B. als Fallschirmjäger, Kraftfahrer/in, Flugabwehrraketensoldat/in, Gebirgs-jäger oder Panzergrenadier ● in Verwendungen tätig sein, die zivilberufliche Qualifikationen voraussetzen, oder in deren Rahmen eine Qualifikation erworben werden kann, z.B. als Radarelektroniksoldat/in, Raketenelektronikmechaniker/in oder Personalsoldat/in ● technische Geräte und Anlagen bedienen, warten und ggf. reparieren ● als Sanitätssoldat/in oder im Militärmusikdienst tätig sein ● im Stabsdienst auf der Streitkräftebasis arbeiten, z.B. die Planung und Durchführung logistischer Einsätze unterstützen, satellitengestützte Aufklärung nutzen ● regelmäßig Sport betreiben, unter teilweise extremen Witterungsbedingungen tätig sein, bei Übungen und Einsätzen Waffen einsetzen ● gegebenenfalls an Auslandseinsätzen teilnehmen
#> 2177:                                                                                                                               Aufgaben, Tätigkeiten, Kenntnisse und Fertigkeiten, üblicherweise: ● militärspezifische Funktionen erfüllen, z.B. als Fallschirmjäger, Kraftfahrer/in, Flugabwehrraketensoldat/in, Gebirgs-jäger oder Panzergrenadier ● in Verwendungen tätig sein, die zivilberufliche Qualifikationen voraussetzen, oder in deren Rahmen eine Qualifikation erworben werden kann, z.B. als Radarelektroniksoldat/in, Raketenelektronikmechaniker/in oder Personalsoldat/in ● technische Geräte und Anlagen bedienen, warten und ggf. reparieren ● als Sanitätssoldat/in oder im Militärmusikdienst tätig sein ● im Stabsdienst auf der Streitkräftebasis arbeiten, z.B. die Planung und Durchführung logistischer Einsätze unterstützen, satellitengestützte Aufklärung nutzen ● regelmäßig Sport betreiben, unter teilweise extremen Witterungsbedingungen tätig sein, bei Übungen und Einsätzen Waffen einsetzen ● gegebenenfalls an Auslandseinsätzen teilnehmen
#>                                                                                                                                                                                                                                                                                                                                                                                                                      Umfasst ferner
#>    1:                                                                                                                                                                                                                                                                                          Die Systematikposition umfasst folgende Unterpositionen: 11 Land-, Forst- und Tierwirtschaftsberufe 12 Gartenbauberufe und Floristik
#>    2:                                                                                                                                                                                                           Die Systematikposition umfasst folgende Unterpositionen: 111 Landwirtschaft 112 Tierwirtschaft 113 Pferdewirtschaft 114 Fischwirtschaft 115 Tierpflege 116 Weinbau 117 Forst- und Jagdwirtschaft, Landschaftspflege
#>    3:                                            Die Systematikposition umfasst folgende Unterpositionen: 1110 Berufe in der Landwirtschaft (ohne Spezialisierung) 1111 Berufe in der Landtechnik 1112 Landwirtschaftliche Sachverständige 1113 Berufe im landwirtschaftlich-technischen Laboratorium 1118 Berufe in der Landwirtschaft (sonstige spezifische Tätigkeitsangabe) 1119 Aufsichts- und Führungskräfte – Landwirtschaft
#>    4: Die Systematikposition umfasst folgende Unterpositionen: 11101Berufe in der Landwirtschaft (ohne Spezialisierung) – Helfer- /Anlerntätigkeiten 11102Berufe in der Landwirtschaft (ohne Spezialisierung) – fachlich ausgerichtete Tätigkeiten 11103Berufe in der Landwirtschaft (ohne Spezialisierung) – komplexe Spezialistentätigkeiten 11104Berufe in der Landwirtschaft (ohne Spezialisierung) – hoch komplexe Tätigkeiten
#>    5:                                                                                                                                                                                                                                                                                                                 Zugeordnete Berufe (Beispiele): Landwirtschaftliche/r Helfer/in Erntehelfer/in Feldarbeiter/in Hofarbeiter/in
#>   ---                                                                                                                                                                                                                                                                                                                                                                                                                              
#> 2173:                                                                                                                                                                                                                                                                                                                                   Die Systematikposition umfasst folgende Unterpositionen: 01302 Unteroffiziere ohne Portepee
#> 2174:                                                                                                                                                                                                                                                           Zugeordnete Berufe (Beispiele): Fahnenjunker Stabsunteroffizier Fachunteroffizier  Fachunteroffizier – Geoinformationsdienst Fachunteroffizier – Militärmusikdienst
#> 2175:                                                                                                                                                                                                                                                                                                       Die Systematikposition umfasst folgende Unterpositionen: 0140 Angehörige der regulären Streitkräfte in sonstigen Rängen
#> 2176:                                                                                                                                                                                                                                                                                                      Die Systematikposition umfasst folgende Unterpositionen: 01402 Angehörige der regulären Streitkräfte in sonstigen Rängen
#> 2177:                                                                                                                                                                                                                                                                     Zugeordnete Berufe (Beispiele): Bundeswehrsoldat Berufssoldat Gefreiter Hauptgefreiter Oberstabsgefreiter Zeitsoldat Fallschirmjäger Matrose (Bundeswehr)
#>                                                                                                                                                                                                                                                                                                                                                                        Ausschlüsse
#>    1:                                                                                                                                                                                                                                                                                                                                                                             
#>    2:                                                                                                                                                                                                                                                                                                                                                                             
#>    3:                                                                                                                                                                                                                       Nicht einzubeziehende Positionen: 112 Tierwirtschaft 114 Fischwirtschaft 115 Tierpflege 117 Forst- und Jagdwirtschaft, Landschaftspflege 121 Gartenbau
#>    4: Nicht einzubeziehende Positionen: 1121Berufe in der Nutztierhaltung (außer Geflügelhaltung) 1130Berufe in der Pferdewirtschaft (ohne Spezialisierung) 1151Berufe in der Nutztierpflege 1160Berufe im Weinbau 1171Berufe in der Forstwirtschaft 1212Berufe in Baumschule, Staudengärtnerei und Zierpflanzenbau 5251Führer/innen von land- und forstwirtschaftlichen Maschinen
#>    5:                                                                                                                                                                                                                       Nicht einzubeziehende Berufe: Tierwirtschaftshelfer/in (11211) Fischereihelfer/in (11401) Forstwirtschaftshelfer/in (11711) Gartenbauhelfer/in (12101)
#>   ---                                                                                                                                                                                                                                                                                                                                                                             
#> 2173:                                                                                                                                                                                                                             Nicht einzubeziehende Positionen: 0110 Offiziere 0120 Unteroffiziere mit Portepee 0140 Angehörige der regulären Streitkräfte in sonstigen Rängen
#> 2174:                                                                                                                                                                                                                                                           Nicht einzubeziehende Berufe: Fähnrich (01203) Stabsfeldwebel (01203) Oberstabsfeldwebel (01203) Feldwebel (01203)
#> 2175:                                                                                                                                                                                                                                                             Nicht einzubeziehende Positionen: 011 Offiziere 012 Unteroffiziere mit Portepee 013 Unteroffiziere ohne Portepee
#> 2176:                                                                                                                                                                                                                                                          Nicht einzubeziehende Positionen: 0110 Offiziere 0120 Unteroffiziere mit Portepee 0130 Unteroffiziere ohne Portepee
#> 2177:                                                                                                                                                                                                                                                                                                                                                                             
#>       Maßeinheit
#>    1:         NA
#>    2:         NA
#>    3:         NA
#>    4:         NA
#>    5:         NA
#>   ---           
#> 2173:         NA
#> 2174:         NA
#> 2175:         NA
#> 2176:         NA
#> 2177:         NA