Show a page with a text field where free text can be entered.

Show a page with a text field where free text can be entered.

Usage

page_freetext(
  page_id,
  question_text = "Please enter your answer in the box below",
  is_interview = FALSE,
  no_answer_checkbox = TRUE,
  next_button = TRUE,
  previous_button = TRUE,
  trigger_next_on_enter = TRUE,
  render_question_text = TRUE,
  run_before = NULL,
  run_after = NULL,
  ...
)

Arguments

page_id

A unique string identifiying this page. Used to store data.

question_text

The question / text to display. This can be either a string, which will simply be displayed or a function to dynamically determine the question_text.

is_interview

Should the page show slightly different / additional instructions and answer options for an interview that is conducted by another person? Defaults to FALSE.

no_answer_checkbox

Whether to provide a checkbox to denote that no answer has been provided.

next_button

Whether to show the button to navigate to the next page? Defaults to TRUE.

previous_button

Whether to show the button to navigate to the preivous page? Defaults to TRUE.

trigger_next_on_enter

Whether the next button is triggered when one presses enter. Defaults to TRUE. There are known issues with IE11.

render_question_text

Whether the question text should be displayed? Only set this to FALSE, if you wish to change the rendering of the question_text by e.g. using render_before. Defaults to TRUE.

run_before

Similar to run_before in new_page(), passed explicitly here as this page adds some of its own code to run_before.

run_after

Similar to run_after in new_page(), passed explicitly here as this page adds some of its own code to run_after.

...

Other parametrs are passed on to new_page()

Value

A page object.

See also

Examples

if (FALSE) { # \dontrun{
page_freetext(
  "test_page_freetext",
  question_text = "Hello there! Please fill in your name below:",
  no_answer_checkbox = TRUE
)
} # }