Make suggestions using similarity based reasoning.

The Algorithm used here corresponds to Algorithm #10 in (Schierholz, 2019). Note: This function should not be used directly, but rather as a step / algorithm in get_job_suggestions.

Usage

algo_similarity_based_reasoning(
  text_processed,
  sim_name = "wordwise",
  probabilities = occupationMeasurement::pretrained_models$similarity_based_reasoning,
  ...
)

Arguments

text_processed

The processed user input. Will be provided by get_job_suggestions.

sim_name

Which similarity measure to use. Possible values are "wordwise" or "substring".

probabilities

Trained probabilities to be used, defaults to the one bundled with the package. See pretrained_models. This pretrained model always predicts a 5-digit code from the 2010 German Classification of Occupations, with some exceptions: -0004 stands for 'Not precise enough/uncodable', -0006 stands for 'Multiple Jobs', -0012 stands for 'Blue-collar workers', -0019 stands for 'Volunteer/Social Service', and -0030 stands for 'Student assistant'.

...

Additional arguments may be passed from get_job_suggestions(), but will be ignored in this function.

Value

A data.table with suggestions or NULL if no suggestions were found.

References

Schierholz, M. (2019). New Methods for Job and Occupation Classification (Ph.D. Thesis). University of Mannheim.

See also

Examples

# Use with default settings
get_job_suggestions(
  "Arzt",
  steps = list(
    simbased_default = list(
      algorithm = algo_similarity_based_reasoning
    )
  )
)
#>    auxco_id input_text      score                                        title
#> 1:     7028       Arzt 0.69811730                       Allgemeinarzt /-ärztin
#> 2:     7029       Arzt 0.05000000 Facharzt/-ärztin – Kinder- und Jugendmedizin
#> 3:     7030       Arzt 0.04799465                                 Internist/in
#> 4:     7031       Arzt 0.04223624                                   Chirurg/in
#> 5:     7032       Arzt 0.03263888                               Anästhesist/in
#>                                                                                                                                                         task
#> 1:                      Untersuchung von Patienten und Diagnose von Krankheiten sowie Festlegung geeigneter Therapiemaßnahmen oder Überweisung zu Fachärzten
#> 2: Untersuchung von minderjährigen Patienten und Diagnose von Kinderkrankheiten sowie Festlegung geeigneter Therapiemaßnahmen oder Überweisung zu Fachärzten
#> 3:                                                                                               Diagnose und Behandlung von Erkrankungen der inneren Organe
#> 4:                                                                                                                 Durchführung von chirurgischen Eingriffen
#> 5:                                                           Durchführung von Anästhesien und Aufrechterhaltung der vitalen Funktionen während der Operation
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                       task_description
#> 1:                                                                                                                                                                                                                                      z.B. Krankheitsbild ermitteln; Diagnose stellen; Röntgenuntersuchungen und andere Diagnoseverfahren veranlassen; Medikamente und Physiotherapie verschreiben; Patienten über Therapiemöglichkeiten informieren; ärztliche Gutachten anfertigen
#> 2:                                                                                                                                                    z.B. Labor- und Röntgenuntersuchungen veranlassen; Krankheitsbilder ermitteln; Eltern über Therapiemöglichkeiten informieren; Medikamente und andere Behandlungsmaßnahmen bei Kinderkrankheiten, Erkältungskrankheiten, Epilepsie oder ADHS verschreiben; ärztliche Gutachten verfassen; über Vorsorgeuntersuchungen informieren
#> 3: z.B. Krankheitsbild ermitteln; Ultraschall-, Röntgen-,EKG -, EEG -, CTG-, Endoskopie-Untersuchungen durchführen; Erkrankungen der Lunge, des Herzens und Kreislaufs, der Verdauungsorgane, der Nieren, des Blutes und des Stoffwechsels behandeln; Infektionskrankheiten therapieren; über Therapmöglichkeiten informieren; Therapieverlauf kontrollieren; über Vorsorgeuntersuchungen informieren; Medikamente verschreiben; ärztliche Gutachten anfertigen; Befunde dokumentieren
#> 4:                                                                                                                                                                                             z.B. Labor-, Röntgen-, Ultraschalluntersuchungen veranlassen; Operationen durchführen; unfallchirurgische Eingriffe tätigen; minimalinvasive Eingriffe vornehmen; therapeutische Maßnahmen für Rehabilitanden veranlassen; Arztberichte und Gutachten anfertigen; Befunde dokumentieren
#> 5:                                                                                                                                                                          z.B. Schmerzanalyse durchführen; Katheter und Sonden legen; Narkose und künstliche Beatmung durchführen; lebenswichtige Körperfunktionen kontrollieren; akute Notfälle erkennen und behandeln; lebensrettende Wiederbelebungsmaßnahmen durchführen; palliativmedizinisch zu versorgende Patienten betreuen
#>                                         kldb_title_short
#> 1:                Ärzte/Ärztinnen (ohne Spezialisierung)
#> 2: Fachärzte/-ärztinnen in der Kinder- und Jugendmedizin
#> 3:           Fachärzte/-ärztinnen in der Inneren Medizin
#> 4:                 Fachärzte/-ärztinnen in der Chirurgie
#> 5:           Fachärzte/-ärztinnen in der Anästhesiologie
#>                                                                                                                   notes
#> 1: Nachfrage nach Kinder- u. Jugendmedizin hinzufügen? Hier wäre mit 81414 eine genauere Zuordnung möglich (2023-07-26)
#> 2:                                                                                                                     
#> 3:                                                                                                                     
#> 4:                                                                                                                     
#> 5:                                                                                                                     
#>    has_followup_questions
#> 1:                   TRUE
#> 2:                  FALSE
#> 3:                  FALSE
#> 4:                  FALSE
#> 5:                  FALSE

# Use with substring similarity
get_job_suggestions(
  "Arzt",
  steps = list(
    simbased_substring = list(
      algorithm = algo_similarity_based_reasoning,
      parameters = list(
        sim_name = "substring"
      )
    )
  )
)
#>    auxco_id input_text      score                  title
#> 1:     7001       Arzt 0.35804461          Arzthelfer/in
#> 2:     7002       Arzt 0.22836029      Zahnarzthelfer/in
#> 3:     7028       Arzt 0.19479173 Allgemeinarzt /-ärztin
#> 4:     7003       Arzt 0.08253183        Arztsekretär/in
#> 5:     7034       Arzt 0.05222537      Zahnarzt/ -ärztin
#>                                                                                                                                    task
#> 1:                                                Unterstützung von Ärzten bei der  Untersuchung, Behandlung und Beratung von Patienten
#> 2:                                            Unterstützung von Zahnärzten bei der  Untersuchung, Behandlung und Beratung von Patienten
#> 3: Untersuchung von Patienten und Diagnose von Krankheiten sowie Festlegung geeigneter Therapiemaßnahmen oder Überweisung zu Fachärzten
#> 4:                                                           Durchführung von Sekretariatstätigkeiten in Krankenhäusern oder Arztpraxen
#> 5:                                                                       Diagnose und Behandlung von Zahn-, Mund- und Kieferkrankheiten
#>                                                                                                                                                                                                                                                                                                                                                                             task_description
#> 1:                                                                                                                                                                                                                                                        z.B. Verbände anlegen; Proben für Laboruntersuchungen nehmen; medizinische Instrumente vorbereiten und warten; Patienten empfangen
#> 2:                                                                                                                                                                         z.B. Füllungen oder Abdrucksmassen anrühren; Röntgenaufnahmen anfertigen; zahnmedizinische Materialien und Instrumente vorbereiten; Patienten vorbeugende Maßnahmen der Mundhygiene erklären; Patienten empfangen
#> 3:                                                                                                                                            z.B. Krankheitsbild ermitteln; Diagnose stellen; Röntgenuntersuchungen und andere Diagnoseverfahren veranlassen; Medikamente und Physiotherapie verschreiben; Patienten über Therapiemöglichkeiten informieren; ärztliche Gutachten anfertigen
#> 4:                                                                                                                                                   z.B. Patienten empfangen; Korrespondenz mit Patienten und Krankenkassen; Termine vereinbaren und Sprechstunden planen; Versicherungsformulare ausfüllen; Atteste oder Krankenberichte verfassen; Leistungen mit Krankenkassen abrechnen
#> 5: z.B. Zahnfüllungen anfertigen oder Inlays einsetzen; erkrankte Zähne aufbohren; Zähne überkronen; Wurzelkanäle behandeln; Beläge und entzündetes Gewebe aus Zahnfleischtaschen entfernen; über Zahnersatz und Zahnersatzpflege informieren und beraten; abgebrochene Zahnwurzeln entfernen; Weisheits- oder andere erkrankte Zähne extrahieren; Zahnspangen in Auftrag geben und anpassen
#>                                            kldb_title_short
#> 1:                             Medizinische Fachangestellte
#> 2:                         Zahnmedizinische Fachangestellte
#> 3:                   Ärzte/Ärztinnen (ohne Spezialisierung)
#> 4:               Verwaltung im Sozial- und Gesundheitswesen
#> 5: Zahnärzte/-ärztinnen und Kieferorthopäden/-orthopädinnen
#>                                                                                                                   notes
#> 1:                                                                                                                     
#> 2:                                                                                                                     
#> 3: Nachfrage nach Kinder- u. Jugendmedizin hinzufügen? Hier wäre mit 81414 eine genauere Zuordnung möglich (2023-07-26)
#> 4:                                                                                                                     
#> 5:                                                                                                                     
#>    has_followup_questions
#> 1:                   TRUE
#> 2:                   TRUE
#> 3:                   TRUE
#> 4:                   TRUE
#> 5:                  FALSE

# Comparison of algo_similarity_based_reasoning() with get_job_suggestions()

# Example of using algo_similarity_based_reasoning() directly. Not recommended.
if (interactive()) {
  algo_similarity_based_reasoning(
    preprocess_string("Arzt"),
    sim_name = "wordwise"
  )[order(score, decreasing = TRUE)]
}

# Same output as before, but the function is more adaptable.
if (interactive()) {
  get_job_suggestions(
    "Arzt",
    suggestion_type = "kldb-2010",
    num_suggestions = 1500,
    steps = list(
      simbased_default = list(
        algorithm = algo_similarity_based_reasoning,
        parameters = list(
          sim_name = "wordwise"
        )
      )
    )
  )[, list(kldb_id, score, sim_name, kldb_id_title = title)]
}