Make coding suggestions based on a user's open-ended text input.

Given a text input, find up to num_suggestions possible occupation categories.

Usage

get_job_suggestions(
  text,
  suggestion_type = "auxco-1.2.x",
  num_suggestions = 5,
  suggestion_type_options = list(),
  aggregate_score_threshold = 0.02,
  item_score_threshold = 0,
  distinctions = TRUE,
  steps = list(simbased_wordwise = list(algorithm = algo_similarity_based_reasoning,
    parameters = list(sim_name = "wordwise")), simbased_substring = list(algorithm =
    algo_similarity_based_reasoning, parameters = list(sim_name = "substring"))),
  include_general_id = FALSE
)

Arguments

text

The raw text input from the user.

suggestion_type

Which type of suggestion to use / provide. Possible options are "auxco-1.2.x" and "kldb-2010".

num_suggestions

The maximum number of suggestions to show. This is an upper bound and less suggestions may be returned. Defaults to 5.

suggestion_type_options

A list with options for generating suggestions. Supported options: - datasets: Pass specific datasets to be used whenn adding information to predictions e.g. use a specific version of the kldb or auxco. Supported datasets are: "auxco-1.2.x", "kldb-2010". By default the datasets bundled with this package are used.

aggregate_score_threshold

A single value or named list of thresholds between 0 and 1. If it is a list, each entry should correspond to one of the steps. If it is a single value, it will apply to all steps. Results from that step will only be returned if the sum of their scores is equal to or greater than the specified threshold. With a aggregate_score_threshold of 0 results will always be returned (if there are any).

item_score_threshold

A threshold between 0 and 1 (usually very small, default 0). Results from any step will only be returned if they are greater than the specified threshold. Allows the removal of highly implausible suggestions.

distinctions

Whether or not to add additional distinctions to similar occupational categories to the source code. Defaults to TRUE.

steps

A list with the algorithms to use and their parameters. Each entry of the list should contain a nested list with two entries: algorithm (the algorithm's function itself) and parameters (the parameters to pass onto the algorithm). Each algorithm will also always have access to a default set of three parameters:

text_processed: The input text after preprocessing

suggestion_type: Which type of suggestion to output

num_suggestions: How many suggestions shall be returned These parameters must not be specified manually and will be provided automatically instead. Defaults to:

list(
  # try similarity "one word at most 1 letter different" first
  list(
    algorithm = algo_similarity_based_reasoning,
    parameters = list(
      sim_name = "wordwise",
      min_aggregate_prob = 0.535
    )
  ),
  # since everything else failed, try "substring" similarity
  list(
    algorithm = algo_similarity_based_reasoning,
    parameters = list(
      sim_name = "substring",
      min_aggregate_prob = 0.02
    )
  )
)

include_general_id

Whether a general column, called "id" should always be returned. This will automatically contain the appropriate id for different suggestion_types i.e. for "auxco-1-2.x" it will contain the same data as the column "auxco_id".

Value

A data.table with suggestions or NULL if no suggestions were found.

Details

The procedure implemented here is, roughly speaking, as follows:

Predict categories from KldB 2010, including their scores. The first algorithm mentioned in steps is used (default: algo_similarity_based_reasoning()).

Convert the predicted KldB 2010 categories to suggestion_type (default: auxco-1.2.x, an n:m mapping, scores are mapped accordingly.). See internal function convert_suggestions() for details.

Remove predicted categories if their score is below item_score_threshold and only keep the num_suggestions top-ranked suggestions.

Start anew, trying the next algorithm in steps, if the the top-ranked suggestions have a low chance to be correct. (Technically, this happens if the summed score of the num_suggestions top-ranked suggestions is below aggregate_score_threshold.)

If suggestion_type == "auxco-1.2.x" and distinctions == TRUE, insert additional and (highly) similar categories or replace existing ones. See internal function add_distinctions_auxco(). Reorder and keep only the num_suggestions top-ranked suggestions. Auxco categories which were added during this step can be identified by their scores: It equals 0.05 for categories with high similarity and 0.005 for categories with medium similarity.

Examples

get_job_suggestions("Koch")
#>    auxco_id input_text      score                             title
#> 1:     9079       Koch 0.86541106                       Koch/Köchin
#> 2:     9088       Koch 0.10094970                     Küchenchef/in
#> 3:     9078       Koch 0.06189197                       Küchenhilfe
#> 4:     9065       Koch 0.05000000               Fischverarbeiter/in
#> 5:     6074       Koch 0.05000000 Fast-Food- und Imbisskoch/-köchin
#>                                                                                          task
#> 1: Zubereitung von Speisen in Restaurants, Hotels und sonstigen gastronomischen Einrichtungen
#> 2:     Führungsaufgaben mit Personalverantwortung in einer Hotel-, Restaurant- oder Großküche
#> 3:                   Ausführung von Hilfsarbeiten in einer Hotel-, Restaurant- oder Großküche
#> 4:           Sortierung, Reinigung und Ausnehmen frischer Fische und deren Weiterverarbeitung
#> 5:                                    Zubereitung von einfachen Speisen in Schnellrestaurants
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               task_description
#> 1:                                                                                                                                                                                                                                                                                                                   z.B. Lebensmittel einkaufen; Lebensmittellieferungen annehmen und prüfen; Waren fachgerecht einlagern und Lagerbestände kontrollieren; Speisekarten und Speisepläne erstellen; Gerichte kreieren; kalte und warme Speisen vorbereiten und zubereiten; Zutaten nach Rezept und persönlichem Urteilsvermögen verwenden; Temperatur von Öfen, Grills, Brätern und anderen Kochgeräten einstellen; Speisen anrichten, optisch ansprechend garnieren und in folgerichtiger Reihenfolge ausgeben; Arbeitsabläufe in der Küche planen; rechtzeitige Zubereitung der Speisen in der richtigen Reihenfolge organisieren; im Veranstaltungsservice Kundenwünsche ermitteln, Speisen zubereiten und an den Veranstaltungsort liefern
#> 2: Führungsaufgaben mit Personalverantwortung bedeutet, dass untergeordnete Arbeitskräfte beaufsichtigt und angeleitet werden. Weitere Aufgaben sind z.B. neue Rezepte kreieren und dokumentieren; Speisepläne und Speisekarten erstellen; Frischprodukte einkaufen; Preise kalkulieren und Ausgabevolumen überwachen; Betriebsabläufe in der Küche planen, koordinieren und steuern; Anweisungen und Aufgaben verteilen; Einhaltung von Kostenplänen und Terminen kontrollieren; Qualitätskontrollen durchführen und Abläufe optimieren; Schulungs- und Weiterbildungsmaßnahmen für Mitarbeiter organisieren und den betrieblichen Teil der Ausbildung organisieren und durchführen; Personalentscheidungen treffen; betriebliche Vorschriften über Hygiene und Lebensmittelsicherheit erklären und durchsetzen; Gesichtspunkte des Natur- und Umweltschutzes beachten und z.B. regionale Produkte verwenden; im Veranstaltungsservice Kundenwünsche ermitteln, Angebote ausarbeiten, Speisen zubereiten und an den Veranstaltungsort liefern
#> 3:                                                                                                                                                                                                                                                                                                                                                                                             Hilfsarbeiten sind einfache Aufgaben, für die keine abgeschlossene Berufsausbildung erforderlich ist, z.B. Gemüse, Salat und Früchte waschen und schneiden; Köche unterstützen und z.B. Zutaten abmessen und mischen; einfache Speisen nach Vorgabe vorbereiten, erwärmen, toasten und anrichten; beim Zubereiten von Marinaden, Soßen und Dressings mithelfen; Küchenmaschinen und -geräte bedienen; Lebensmittel aus Vorrats-, Kühlräumen holen und bereitstellen; verderbliche Waren in Kühlräumen lagern und kontrollieren; beim Aussortieren und Entsorgen verfallener Ware mitarbeiten; Aufräum-, Spül- und Reinigungsarbeiten übernehmen
#> 4:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           z.B. Fische in Verarbeitungsanlagen sortieren, reinigen, ausnehmen und häuten; Fische filetieren und durch z.B. Würzen, Braten oder Räuchern weiterverarbeiten; Produkte konservieren oder tiefkühlen, portionieren und verpacken; Produkte während des Prozesses auf Geschmack, Konsistenz und Geruch prüfen; Produktionsanlagen Anlagen zur Fischverarbeitung reinigen, warten und pflegen; die Einhaltung von Qualitätsstandards sicherstellen; Konsumenten über Herkunft, Qualität und Zubereitung der Fischereiwaren beraten
#> 5:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               z.B. in der Küche beim Vor- und Zubereiten einfacher Speisen, wie Sandwiches, Hamburger, Pizzas, Pommes und Salaten mithelfen; Nahrungsmittel waschen, schneiden und abwiegen; große Einfachkochgeräte, wie Grills, Friteusen oder Backblechen, bedienen; fertige Speisen aufwärmen; Kochflächen und Küchenutensilien reinigen; Aufnahme von Bestelllungen; Hygiene am Arbeitsplatz einhalten
#>                             kldb_title_short has_followup_questions
#> 1:                           Köche/Köchinnen                   TRUE
#> 2:        Speisenzubereitung (Führungskraft)                   TRUE
#> 3:                           Köche/Köchinnen                  FALSE
#> 4:                         Fischverarbeitung                   TRUE
#> 5: Gastronomieservice (ohne Spezialisierung)                  FALSE

get_job_suggestions("Schlosser")
#>    auxco_id input_text      score                         title
#> 1:     2034  Schlosser 0.68830113                Metallbauer/in
#> 2:     1846  Schlosser 0.05000000         Metallbautechniker/in
#> 3:     2040  Schlosser 0.05000000            Metallbauhelfer/in
#> 4:     2076  Schlosser 0.12068049 Kraftfahrzeugmechatroniker/in
#> 5:     2059  Schlosser 0.05945473         Maschinenschlosser/in
#>                                                                                                      task
#> 1:                                     Montage und Aufbereitung von Metallteilen und Metallkonstruktionen
#> 2:                                           Planung, Koordination und Überwachung von Metallbauprojekten
#> 3: Ausführung von Hilfsarbeiten bei der Bearbeitung, Montage und Endverarbeitung von Metallkonstruktionen
#> 4:                                                              Wartung und Reparatur von Kraftfahrzeugen
#> 5:                      Aufbau und Montage von Maschinen und Anlagen sowie die Herstellung ihrer Bauteile
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  task_description
#> 1:                                                                                                 z.B. gefertigte Bauteile und fertige Bau- und Zubehörteile zu festen und beweglichen Metallkonstruktionen, z.B. Hallen, Schiffen, Rolltreppen, Toren oder Gittern, montieren; Handwerkzeuge benutzen und nummerisch oder konventionell gesteuerte Maschinen einrichten und bedienen; Stahlträger und Bleche genau nach Maß schneiden, abkanten oder biegen; Bohrungen und Schweißfugen anbringen; Sanierungsarbeiten vornehmen
#> 2:            z.B. Leichtmetallbauteile und -konstruktionen entwerfen; Konstruktionszeichnungen mittels rechnergesteuerten Zeichenanlagen anfertigen; Materialbedarf ermitteln; Herstellungsprozesse in technischer und terminlicher Hinsicht planen und überwachen; Einrichtung von Montagebaustellen planen; Qualitätskontrollen an Metallerzeugnissen durchführen; neue Theorien und Methoden im Metallbau erforschen und entwickeln; Kontrollsysteme zur Sicherstellung einer effizienten Funktion der Konstruktion festlegen
#> 3: Hilfsarbeiten sind einfache Aufgaben, für die keine abgeschlossene Berufsausbildung erforderlich ist, z.B. Fachkräfte bei Bohr-, Stanz- und Schneidarbeiten für den Bau von Metallteilen unterstützen; bei der manuellen bzw. maschinellen Bearbeitung und Verformung von Stahl- und Nichteisenblechen zuarbeiten; Verschraubungen und andere Verbindungen auf Montagebaustellen herstellen sowie anfallende Transportarbeiten verrichten; beim Reinigen und Instandhalten von Werkzeugen, Maschinen und Arbeitsgeräten helfen
#> 4:                                                                                                                                    z.B. Störungen an Systemen und Bauteilen von Fahrzeugen diagnostizieren und dazu ggf. Computer verwenden; elektrische, elektronische, mechanische und pneumatische Systeme in Fahrzeugen nach amtlichen Vorgaben prüfen; Personenkraftwagen mit Zusatzeinrichtungen, Sonderausstattungen und Zubehörteilen ausrüsten; Motoren, Bremsen, Getriebe und sonstige Teile von Fahrzeugen einbauen
#> 5:                                                                                                                                                                                      z.B. Produktionsmaschinen und Fertigungsanlagen, wie auch Aufzüge oder Förderbänder einstellen; Bauteile zu Maschinen montieren und deren Funktion prüfen; benötigte Bauteile durch Drehen, Fräsen, Bohren und Schleifen bearbeiten; elektronische, pneumatische und hydraulische Steuerungen installieren; Montagezeichnungen anfertigen
#>                                            kldb_title_short
#> 1:                                                Metallbau
#> 2:                                                Metallbau
#> 3:                                                Metallbau
#> 4:                                     Kraftfahrzeugtechnik
#> 5: Maschinenbau- und Betriebstechnik (ohne Spezialisierung)
#>    has_followup_questions
#> 1:                  FALSE
#> 2:                   TRUE
#> 3:                  FALSE
#> 4:                  FALSE
#> 5:                  FALSE