Show a page with a text field where free text can be entered.

Show a page with a text field where free text can be entered.

Usage

page_freetext(
  page_id,
  question_text = "Please enter your answer in the box below",
  is_interview = FALSE,
  no_answer_checkbox = TRUE,
  next_button = TRUE,
  previous_button = TRUE,
  trigger_next_on_enter = TRUE,
  render_question_text = TRUE,
  run_before = NULL,
  run_after = NULL,
  ...
)

Arguments

page_id

A unique string identifiying this page. Used to store data.

question_text

The question / text to display. This can be either a string, which will simply be displayed or a function to dynamically determine the question_text.

is_interview

Should the page show slightly different / additional instructions and answer options for an interview that is conducted by another person? Defaults to FALSE.

no_answer_checkbox

Whether to provide a checkbox to denote that no answer has been provided.

next_button

Whether to show the button to navigate to the next page? Defaults to TRUE.

previous_button

Whether to show the button to navigate to the preivous page? Defaults to TRUE.

trigger_next_on_enter

Whether the next button is triggered when one presses enter. Defaults to TRUE. There are known issues with IE11.

render_question_text

Whether the question text should be displayed? Only set this to FALSE, if you wish to change the rendering of the question_text by e.g. using render_before. Defaults to TRUE.

run_before

Similar to run_before in new_page(), passed explicitly here as this page adds some of its own code to run_before.

run_after

Similar to run_after in new_page(), passed explicitly here as this page adds some of its own code to run_after.

...

Other parametrs are passed on to new_page()

Value

A page object.

See also

Examples

page_freetext(
  "test_page_freetext",
  question_text = "Hello there! Please fill in your name below:",
  no_answer_checkbox = TRUE
)
#> $page_id
#> [1] "test_page_freetext"
#> 
#> $condition
#> NULL
#> 
#> $run_before
#> function (session, page, ...) 
#> {
#>     if (is.function(question_text)) {
#>         actual_question_text <- question_text(session, page, 
#>             ...)
#>     }
#>     else {
#>         actual_question_text <- question_text
#>     }
#>     set_item_data(session = session, page_id = page$page_id, 
#>         question_text = actual_question_text)
#>     if (no_answer_checkbox) {
#>         set_item_data(session = session, page_id = page$page_id, 
#>             item_id = "no_answer", question_text = actual_question_text)
#>     }
#>     if (!is.null(run_before)) {
#>         extra_run_before_output <- run_before(session, page, 
#>             ...)
#>     }
#>     else {
#>         extra_run_before_output <- list()
#>     }
#>     utils::modifyList(list(question_text = actual_question_text), 
#>         extra_run_before_output)
#> }
#> <bytecode: 0x5634d5d7a030>
#> <environment: 0x5634d46ee9b8>
#> 
#> $render_before
#> NULL
#> 
#> $render
#> function (session, page, run_before_output, ...) 
#> {
#>     list(if (render_question_text) p(run_before_output$question_text), 
#>         if (trigger_next_on_enter) {
#>             tags$script(paste0("var $nextButton = $(\"#nextButton\");\n                      $(\"#", 
#>                 page$page_id, "_text\").on(\"keyup\", function(e) {\n                        if(e.keyCode == 13){\n                          $nextButton.click();\n                        }\n                      });"))
#>         }, textInput(paste0(page$page_id, "_text"), NULL, value = get_item_data(session = session, 
#>             page_id = page$page_id, key = "response_text"), width = "80%"), 
#>         br(), if (no_answer_checkbox) {
#>             checkboxInput(paste0(page$page_id, "_chk_no_answer"), 
#>                 label = if (is_interview) {
#>                   p(class = "interviewer", "*** Keine Angabe")
#>                 } else {
#>                   p("Keine Angabe")
#>                 }, value = get_item_data(session = session, page_id = page$page_id, 
#>                   item_id = "no_answer", key = "response_id", 
#>                   default = FALSE) == "no_answer_checked", width = "100%")
#>         }, if (previous_button) button_previous(), if (next_button) button_next())
#> }
#> <bytecode: 0x5634d5d788c8>
#> <environment: 0x5634d46ee9b8>
#> 
#> $render_after
#> NULL
#> 
#> $run_after
#> function (session, page, input, ...) 
#> {
#>     set_item_data(session = session, page_id = page$page_id, 
#>         response_text = if (is.null(input[[paste0(page$page_id, 
#>             "_text")]])) 
#>             ""
#>         else input[[paste0(page$page_id, "_text")]])
#>     if (no_answer_checkbox) {
#>         no_answer_checked <- if (is.null(input[[paste0(page$page_id, 
#>             "_chk_no_answer")]])) 
#>             FALSE
#>         else input[[paste0(page$page_id, "_chk_no_answer")]]
#>         set_item_data(session = session, page_id = page$page_id, 
#>             item_id = "no_answer", response_id = if (no_answer_checked) 
#>                 "no_answer_checked"
#>             else "no_answer_unchecked", response_text = if (no_answer_checked) 
#>                 "Keine Angabe"
#>             else "")
#>     }
#>     if (!is.null(run_after)) {
#>         run_after(session, page, input, ...)
#>     }
#> }
#> <bytecode: 0x5634d5d7a890>
#> <environment: 0x5634d46ee9b8>
#> 
#> attr(,"class")
#> [1] "page"